move logging callback + dead letter queue to separate files
